feat: separate programs by functions
